@@ -1,6 +1,6 @@
 {
-  "name": "sakuraapi",
-  "version": "0.3.0",
+  "name": "@sakuraapi/api",
+  "version": "0.4.0",
   "description": "A NodeJS API Experiment",
   "main": "index.js",
   "typings": "index.d.ts",
@@ -11,12 +11,22 @@
     "!docs/**/*"
   ],
   "scripts": {
-    "build": "tsc && npm run doc:generate",
-    "doc:generate": "rm -rf docs && typedoc --options typedoc.json index.ts",
+    "build": "tsc",
+    "doc:generate": "rm -rf docs && typedoc --options typedoc.json **/**",
+    "doc:serve": "nodemon --config nodemon.typedoc.json -e ts --exec 'npm run doc:generate && npm run http-server docs/ -- -a localhost -c-1'",
+    "docker:compose-test": "(docker volume prune -f || echo \"skipped docker volume prune\") ; docker-compose up -d --remove-orphans",
+    "http-server": "http-server",
     "jasmine": "jasmine",
+    "lint": "npm run build && npm run lint:code && npm run lint:tests",
+    "lint:code": "npm run tslint -- -e \"node_modules/**/**/*.ts\"  -e \"**/**/*.d.ts\" -e \"**/**/*.spec.ts\" -t stylish **/**/*.ts",
+    "lint:tests": "npm run tslint -- -e \"node_modules/**/**/*.ts\"  -e \"**/**/*.d.ts\" -t stylish --fix -c tslint.spec.json **/**/*.spec.ts",
     "prepublish": "npm test",
-    "test": "npm run build && npm run jasmine",
+    "test": "npm run docker:compose-test && npm run build && npm run jasmine && docker-compose down",
+    "test:debug": "DEBUG=sapi:*,-sapi:*:verbose npm test",
+    "test:verbose": "DEBUG=sapi:* npm test",
+    "test:vverbose": "DEBUG=* npm test",
     "tsc": "tsc",
+    "tslint": "tslint",
     "typedoc": "typedoc"
   },
   "repository": {
@@ -30,7 +40,7 @@
     "server"
   ],
   "author": "Jean-Pierre E. Poveda",
-  "license": "MIT",
+  "license": "BSD-3-Clause",
   "bugs": {
     "url": "https://github.com/sakuraapi/api/issues"
   },
@@ -39,24 +49,30 @@
     "node": ">=7.0.0"
   },
   "devDependencies": {
-    "@types/colors": "^0.6.33",
-    "@types/express": "^4.0.34",
-    "@types/jasmine": "^2.5.38",
-    "@types/lodash": "^4.14.40",
-    "@types/node": "^6.0.51",
-    "@types/supertest": "^1.1.32",
-    "install": "^0.8.2",
-    "jasmine": "^2.5.2",
+    "@types/colors": "^1.1.2",
+    "@types/express": "^4.0.35",
+    "@types/jasmine": "2.5.41",
+    "@types/lodash": "^4.14.61",
+    "@types/mongodb": "^2.1.41",
+    "@types/node": "^7.0.12",
+    "@types/supertest": "^2.0.0",
+    "http-server": "^0.9.0",
+    "jasmine": "^2.5.3",
+    "jasmine-spec-reporter": "^3.2.0",
     "nodemon": "^1.11.0",
-    "npm": "^4.0.2",
-    "supertest": "^2.0.1",
-    "typedoc": "^0.5.1",
-    "typescript": "^2.0.10"
+    "supertest": "^3.0.0",
+    "tslint": "^5.0.0",
+    "typedoc": "^0.5.9",
+    "typescript": "^2.2.2"
   },
   "dependencies": {
+    "bson-ext": "^1.0.5",
+    "co": "^4.6.0",
     "colors": "^1.1.2",
-    "express": "^4.14.0",
-    "lodash": "^4.17.2",
-    "reflect-metadata": "^0.1.8"
+    "debug": "^2.6.3",
+    "express": "^4.15.2",
+    "lodash": "^4.17.4",
+    "mongodb": "^2.2.25",
+    "reflect-metadata": "^0.1.10"
   }
 }
