@@ -1,20 +1,14 @@
-import {Json}    from './json';
+import {Json} from './json';
 import {
   IModel,
   Model
-}                from './model';
+} from './model';
 import {Private} from './private';
 
-describe('@Private', function () {
+describe('@Private', function() {
   @Model()
   class Test implements IModel {
 
-    @Private()
-    private allowOverride = true;
-
-    @Private()
-    private denyOverride = false;
-    
     @Private()
     aPrivateField1: number = 1;
 
@@ -41,63 +35,78 @@ describe('@Private', function () {
 
     aPublicField: number = 777;
 
+    @Private()
+    private allowOverride = true;
+
+    @Private()
+    private denyOverride = false;
+
     private overrideFn() {
       return this.allowOverride;
-    };
+    }
 
     private denyOverrideFn() {
       return this.denyOverride;
     }
   }
 
-  beforeEach(function () {
+  beforeEach(function() {
     this.t = new Test();
   });
 
-  describe('toJson', function () {
-    it('excludes a field decorated with @Private()', function () {
-      let json = this.t.toJson();
-      expect(json.aPrivateField1).toBeUndefined();
+  describe('toJson', function() {
+    it('excludes a field decorated with @Private()', function() {
+      const json = this.t.toJson();
+      expect(json.aPrivateField1)
+        .toBeUndefined();
     });
 
-    it(`includes a field decorated with @Private('a_function_returning_true')`, function () {
-      let json = this.t.toJson();
-      expect(json.aPrivateField2).toBe(2);
+    it(`includes a field decorated with @Private('a_function_returning_true')`, function() {
+      const json = this.t.toJson();
+      expect(json.aPrivateField2)
+        .toBe(2);
     });
 
-    it(`excludes a field decorated with @Private('a_function_returning_false')`, function () {
-      let json = this.t.toJson();
-      expect(json.aPrivateField3).toBeUndefined();
+    it(`excludes a field decorated with @Private('a_function_returning_false')`, function() {
+      const json = this.t.toJson();
+      expect(json.aPrivateField3)
+        .toBeUndefined();
     });
 
-    it(`includes a field decorated with @Private('a_property_that_is_truthy`, function () {
-      let json = this.t.toJson();
-      expect(json.aPrivateField4).toBe(4);
+    it(`includes a field decorated with @Private('a_property_that_is_truthy`, function() {
+      const json = this.t.toJson();
+      expect(json.aPrivateField4)
+        .toBe(4);
     });
 
-    it(`excludes a field decorated with @Private('a_property_that_is_falsy`, function () {
-      let json = this.t.toJson();
-      expect(json.aPrivateField5).toBeUndefined();
+    it(`excludes a field decorated with @Private('a_property_that_is_falsy`, function() {
+      const json = this.t.toJson();
+      expect(json.aPrivateField5)
+        .toBeUndefined();
     });
 
-    it('excludes a field that is decorated with both @Private and @Json with an alias', function () {
-      let json = this.t.toJson();
-      expect(json.aPrivateField6).toBeUndefined();
+    it('excludes a field that is decorated with both @Private and @Json with an alias', function() {
+      const json = this.t.toJson();
+      expect(json.aPrivateField6)
+        .toBeUndefined();
     });
 
-    it('includes a field that is decorated with @Private and has a truth value', function () {
-      let json = this.t.toJson();
-      expect(json.aPrivateField7).toBeUndefined();
+    it('includes a field that is decorated with @Private and has a truth value', function() {
+      const json = this.t.toJson();
+      expect(json.aPrivateField7)
+        .toBeUndefined();
     });
 
-    it('excludes a field that is decorated with @Private and has a falsy value', function () {
-      let json = this.t.toJson();
-      expect(json.aPrivateField8).toBeUndefined();
+    it('excludes a field that is decorated with @Private and has a falsy value', function() {
+      const json = this.t.toJson();
+      expect(json.aPrivateField8)
+        .toBeUndefined();
     });
 
-    it('does not exclude a field that is not decorated with @Private', function () {
-      let json = this.t.toJson();
-      expect(json.aPublicField).toBe(777);
+    it('does not exclude a field that is not decorated with @Private', function() {
+      const json = this.t.toJson();
+      expect(json.aPublicField)
+        .toBe(777);
     });
   });
 });
