@@ -1,23 +1,27 @@
+/* tslint:disable:variable-name */
+
 /**
- * This is an internal helper function used by the SakuraApi framework. It's behavior is not part of the
+ * @warning @internal This is an internal helper function used by the SakuraApi framework. It's behavior is not part of the
  * official API and as a result, that behavior may break without notice.
+ * @internal This is not part of the API contract.
  */
-export function addDefaultInstanceMethods(target: any, functionName: string, fn: (any) => any, options?: any) {
+export function addDefaultInstanceMethods(target: any, functionName: string, fn: (...any) => any, options?: any) {
 
   if (options && options.suppressInjection && options.suppressInjection.indexOf(functionName) > -1) {
     return;
   }
 
   if (!target.prototype[functionName]) {
-    target.prototype[functionName] = fn.bind(target);
+    target.prototype[functionName] = fn;
   }
 }
 
 /**
- * This is an internal helper function used by the SakuraApi framework. It's behavior is not part of the
+ * @warning This is an internal helper function used by the SakuraApi framework. It's behavior is not part of the
  * official API and as a result, that behavior may break without notice.
+ * @internal This is not part of the API contract.
  */
-export function addDefaultStaticMethods(target: any, functionName: string, fn: (any) => any, options?: any) {
+export function addDefaultStaticMethods(target: any, functionName: string, fn: (...any) => any, options?: any) {
 
   if (options && options.suppressInjection && options.suppressInjection.indexOf(functionName) > -1) {
     return;
@@ -27,3 +31,4 @@ export function addDefaultStaticMethods(target: any, functionName: string, fn: (
     target[functionName] = fn;
   }
 }
+/* tslint:enable:variable-name */
