@@ -1,3 +1,3 @@
-module.exports = new function () {
+module.exports = new (function() {
   this.envTs = 777;
-};
+})();
